### Context

Was going to add to this spec so was improving the spec before I added to it. Leaving the camp a bit tidier on passing through.

Specs on this project are a bit on the slow side and it's death by 1000
papercuts, so adding a small improvement.

### Changes proposed in this pull request

* Spec refactor - don't repeat request for status code check
* Spec refactor - split provider spec in two

#### before

```
bundle exec spec/requests/api/v2/providers_spec.rb
...
Finished in 9.67 seconds (files took 3.35 seconds to load)
17 examples, 0 failures
```

#### after

```
bundle exec spec/requests/api/v2/providers/providers_s*
...
Finished in 6.68 seconds (files took 3.1 seconds to load)
13 examples, 0 failures
```

That's a [very unscientifically measured] 30.9% faster :muscle: 

### Guidance to review

Read the individual commits. First commit just splits the file in two.

:eyes:

### Checklist

- [ ] Make sure all information from the Trello card is in here
- [ ] Attach to Trello card
- [x] Rebased master
- [x] Cleaned commit history
- [x] Tested by running locally