@timabell 2020-03-25
Added another cleanup, squashed in